+++ server/errors.rs
@@ -7,6 +7,8 @@ pub enum Error {
     Server(StatusCode, String),
     #[error("{0}")]
     Generation(String),
+    #[error("{0}")]
+    BoxedError(Box<dyn std::error::Error + Send + Sync>),
     // Froms
     #[error("{0}")]
     MongoDB(#[from] mongodb::error::Error),
@@ -26,6 +28,12 @@ pub enum Error {
     TimeConversion(#[from] time::error::ConversionRange),
 }
 
+impl From<Box<dyn std::error::Error + Send + Sync>> for Error {
+    fn from(error: Box<dyn std::error::Error + Send + Sync>) -> Self {
+        Error::BoxedError(error)
+    }
+}
+
 impl IntoResponse for Error {
     fn into_response(self) -> Response {
         let msg = format!("{}", self.to_string());
@@ -40,6 +48,7 @@ impl From<Error> for StatusCode {
         match error {
             Error::Server(c, _) => c,
             Error::Generation(_) => StatusCode::BAD_REQUEST,
+            Error::BoxedError(_) => StatusCode::INTERNAL_SERVER_ERROR,
             _ => StatusCode::INTERNAL_SERVER_ERROR,
         }
     }